+++ src/hyperlight_common/src/flatbuffer_wrappers/mod.rs
@@ -18,4 +18,4 @@ pub mod host_function_definition;
 pub mod host_function_details;
 pub mod util;
 
-pub use codec::{ExternalValueRef, ExternalValueRefs, ExternalValueSink, ExternalValueSource};
+pub use codec::{ExternalValueSink, ExternalValueSource};